+++ lib/OneWire/library.properties
@@ -1,0 +1,10 @@
+name=OneWire
+version=2.3.3
+author=Jim Studt, Tom Pollard, Robin James, Glenn Trewitt, Jason Dangel, Guillermo Lovato, Paul Stoffregen, Scott Roberts, Bertrik Sikken, Mark Tillotson, Ken Butcher, Roger Clark, Love Nystrom
+maintainer=Paul Stoffregen
+sentence=Access 1-wire temperature sensors, memory and other chips.
+paragraph= Mod of Paul Stoffregen code to support ESP32
+category=Communication
+url=http://www.pjrc.com/teensy/td_libs_OneWire.html
+architectures=esp8266,esp32
+
